requirement.txt
python-telegram-bot
google.generativeai
python-dotenv
pdfplumber